add Input component
