+++ App.tsx
@@ -1,28 +1,19 @@
 import { StatusBar } from 'expo-status-bar';
-import { StyleSheet, Text, TextInput, View } from 'react-native';
+import { StyleSheet, Text,  View } from 'react-native';
 import Header from './components/Header';
-import { useState } from 'react';
+import Input from './components/Input';
 
 
 export default function App() {
-  const [text, setText] = useState("");
 
   const appName = "My Awesome App";
-  function updateText(changedText: string) {
-    // update the text state
-    setText(changedText);
-  }
 
   return (
     <View style={styles.container}>
       {/* <Text>Welcome to {appName}</Text> */}
       <StatusBar style="auto" />
-      <Header name={appName}/>
-      {}
-      <TextInput 
-        value={text} 
-        onChangeText={updateText}
-      />
+      <Header name={appName} />
+      <Input />
     </View>
   );
 }
